removed repeated code in query responder
The socket writes were being repeated needlessly in each match branch. This change simplifies this entire block while also reducing the level of scopes. It could be improved further by abstracting the query match into a separate function or even building out a more generic method of filtering and reacting to specific packet patterns. But I chose to leave this for now until the Query API is re-designed a bit.